Add more generic functions

+++ day16/day16_test.go
@@ -3,7 +3,7 @@ package day16
 import "testing"
 
 func TestRegression(t *testing.T) {
-	right := []int64{21071, 3429967441937}
+	right := []int{21071, 3429967441937}
 	left := []interface{}{RunPart1().Value, RunPart2().Value}
 
 	for i, l := range left {
